backend/face_cut/requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
cmake==3.18.4.post1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
easydict==1.9
graphviz==0.8.4
httplib2==0.9.2
idna==2.6
imageio==2.9.0
insightface==0.1.5
joblib==1.0.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.3.3
mxnet==1.7.0.post1
netifaces==0.10.4
networkx==2.5
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
opencv-python==4.5.1.48
packaging==20.8
paho-mqtt==1.5.1
pexpect==4.2.1
Pillow==8.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
pymongo==3.11.2
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.25.1
requests-unixsocket==0.1.5
scikit-build==0.11.1
scikit-image==0.17.2
scikit-learn==0.24.0
scipy==1.5.4
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
threadpoolctl==2.1.0
tifffile==2020.9.3
tqdm==4.56.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
xkit==0.0.0
zope.interface==4.3.2
